- Allow the Spinner to be used with ranges other than simple numbers.
- Range classes know best how to increment and decrement values within the
  range.  This is particularly true when the ranges are not simple numbers.
  Spinner now requests that the Range manager increment/decrement values if
  the manager supports it (provides incrementValue() and pageIncrementValue()
  methods.

Although I've tested this locally, right now it's difficult to provide a
qooxdo sample to demonstrate these features because they requires a custom
Range class, but there's no way to easily add classses to a sample with each
sample in its own file.  Once the samples are refactored into individual
skeleton-based applications, it will be easy to provide a sample for this.

Derrell