Fix column redisplay when a column becomes visible.
The default behavior has been extended such that when a column becomes visible
(and thus a "visibilityChanged" event is generated), the column widths are
recalculated based on the same rule as used for initial display on "appear".

It would be nice if hidden column widths were saved so that the sequence of
(a) adjusting column widths manually, (b) making a column invisible, and (c)
making that column visible again would maintain the manually-adjusted column
widths.  That's a bit much for the Default behavior though.  I expect someone
will create a new behavior which extends the Default one, that saves column
widths.

Derrell