[BUG #ASSIGNED 890]: “Mixed-Content” warning in Internet Explorer working over HTTPS http://bugzilla.qooxdoo.org/show_bug.cgi?id=890
Fixed bug reported by Denis. 

IE has a problem with CSS defined background images delivered via HTTPS. Even
if the paths to the images are defined as relative ones.
To avoid the problem with the "Mixed-Content" warning qooxdoo will in future use 
absolute URL's if the application is delivered via HTTPS in IE.