[BUG #ASSIGNED 1717]: IE looses format styles at empty paragraphs when focus is lost http://bugzilla.qooxdoo.org/show_bug.cgi?id=1717
Solution: check at every "focusOut" event (=blur of the editor) if there is an empty paragraph. If so collect the formatting style of the previous paragraph and memorize it.
If the next "execCommand" is executed without any "keyPress" event before format the current empty paragraph with the collected styles of the previous one.