Add a way to limit the services exposed by handle_request()
For backward compatibility reasons the module will by default expose
all Qooxdoo::Services::* modules on the Perl's search path. This is a
bit liberal, and most applications should probably only make a known
limited subset of services available.  This can be accomplished by
passing a reference to a list of allowed service names or regular
expressions as the third parameter of the handle_request function.

Also improve the test suite coverage.