Added support for a new service method call signature through the "RpcMethodSignatureMode" constant. This constant can be defined in the global_settings.php configuration file (or elsewehre as long as it defined before including the RpcPhp files) and determines how to pass the RPC parameters to the service method. It can take three different values:
1) 'array': The signature used in RpcPhp 1.0, passing all
parameters as an array in the first method argument, and the error
object as the second argument. This is the default mode (if the 
constant is not defined by the user) and the only one that works in PHP4.

2) 'args': Pass the parameters as a normal method call, so that
the method definition can use them as named arguments. This also
improves the ability to document the source code of the service methods.
This will be the standard mode in a future version of the server.
The error object does not need to be passed to the service method any
longer.

3) 'check': Via method introspection, this checks whether the first
argument of the method is called "params". If yes, use the 'array' mode.
If not, use the 'args' mode.