Added password encryption to the access controller and access service.
Before, the password was stored as plain text in the database. Now, it is encrypted with the sha1 algorythm, using a random salt which is prepended to the encrypted password --- see qcl_access_Controller::generateHash().

When you upgrade, manually change the length of the "password" column in the data_User table from 32 to 50 bytes, for some reason the length is not automatically upgraded. 

You will also need to update any password that had been encrypted with a simple md5() function before by replacing it with the plain text password and registering users with the new qcl_access_Controller::register() method.