Updates to RpcConsole
The most significant change here is that I've added a function for checking a
returned error. This allows testing negative, as well as positive conditions,
and the qooxdoo.test.getError test now "passes" by receiving the expected
error.

I also fixed the getParam() test which is supposed to take only a single
parameter (some backends don't support varargs).

Finally, I "normalized" the parameters to getParams() to include the floating
point test that was supposed to be one of the parameters.

Patch has been approved by Christian.

Derrell