Yet again a slight modification of the authentication process:
It is now REQUIRED to pass a session id to the server UNLESS the application exempts explicitly all or specific method calls in the qcl_application_Application::skipAuthentication() method (by returning a True value). In this case the session id is taken from the PHP session which is common to all browser windows.